requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.3.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
blis==0.7.11
    # via thinc
cachetools==5.4.0
    # via streamlit
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   streamlit
    #   typer
cloudpathlib==0.18.1
    # via weasel
confection==0.1.5
    # via
    #   thinc
    #   weasel
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
cymem==2.0.8
    # via
    #   preshed
    #   spacy
    #   thinc
distro==1.9.0
    # via openai
faiss-cpu==1.8.0.post1
    # via -r requirements.in
fonttools==4.53.1
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
    #   spacy
    #   st-copy-to-clipboard
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
langchain-core==0.2.25
    # via langchain-openai
langchain-openai==0.1.19
    # via -r requirements.in
langcodes==3.4.0
    # via spacy
langsmith==0.1.94
    # via langchain-core
language-data==1.2.0
    # via langcodes
marisa-trie==1.2.0
    # via language-data
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.1
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
murmurhash==1.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
neo4j==5.23.0
    # via -r requirements.in
networkx==3.3
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   altair
    #   blis
    #   contourpy
    #   faiss-cpu
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   spacy
    #   streamlit
    #   thinc
openai==1.37.1
    # via
    #   -r requirements.in
    #   langchain-openai
orjson==3.10.6
    # via langsmith
packaging==24.1
    # via
    #   altair
    #   faiss-cpu
    #   langchain-core
    #   matplotlib
    #   spacy
    #   streamlit
    #   thinc
    #   weasel
pandas==2.2.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==10.4.0
    # via
    #   matplotlib
    #   streamlit
preshed==3.0.9
    # via
    #   spacy
    #   thinc
protobuf==5.27.3
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.8.2
    # via
    #   confection
    #   langchain-core
    #   langsmith
    #   openai
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.20.1
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pymupdf==1.24.9
    # via -r requirements.in
pymupdfb==1.24.9
    # via pymupdf
pyparsing==3.1.2
    # via matplotlib
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via
    #   neo4j
    #   pandas
pyyaml==6.0.1
    # via langchain-core
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   langsmith
    #   spacy
    #   streamlit
    #   tiktoken
    #   weasel
rich==13.7.1
    # via
    #   streamlit
    #   typer
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
smart-open==7.0.4
    # via weasel
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
spacy==3.7.5
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
srsly==2.4.8
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
st-copy-to-clipboard==0.1.6
    # via -r requirements.in
streamlit==1.37.0
    # via
    #   -r requirements.in
    #   st-copy-to-clipboard
tenacity==8.5.0
    # via
    #   langchain-core
    #   streamlit
thinc==8.2.5
    # via spacy
tiktoken==0.7.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
tornado==6.4.1
    # via streamlit
tqdm==4.66.4
    # via
    #   -r requirements.in
    #   openai
    #   spacy
typer==0.12.3
    # via
    #   spacy
    #   weasel
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
wasabi==1.1.3
    # via
    #   spacy
    #   thinc
    #   weasel
weasel==0.4.1
    # via spacy
wrapt==1.16.0
    # via smart-open

# The following packages are considered to be unsafe in a requirements file:
# setuptools
